+++ src/resolutionError.ts
@@ -28,7 +28,7 @@ const HandlersByCode = {
     `${params.currencyTicker} is not supported`,
   IncorrectResolverInterface: (params: { method: string }) =>
     `Domain resolver is configured incorrectly for ${params.method}`,
-  RecordNotFound: (params: {recordName: string, domain: string }) =>
+  RecordNotFound: (params: { recordName: string; domain: string }) =>
     `No ${params.recordName} record found for ${params.domain}`,
 };
 
@@ -52,7 +52,9 @@ export default class ResolutionError extends Error {
   constructor(code: ErrorCode, options: ResolutionErrorOptions = {}) {
     const resolutionErrorHandler: ResolutionErrorHandler = HandlersByCode[code];
     const { domain, method, currencyTicker, recordName } = options;
-    super(resolutionErrorHandler({ domain, method, currencyTicker, recordName }));
+    super(
+      resolutionErrorHandler({ domain, method, currencyTicker, recordName }),
+    );
     this.code = code;
     this.domain = domain;
     this.method = method;